requirements.txt
# Core Django
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Authentication
djangorestframework-simplejwt>=5.2.0

# Database
psycopg2-binary>=2.9.0

# Environment variables
python-decouple>=3.8

# Image processing
Pillow>=10.0.0

# API documentation
drf-spectacular>=0.26.0

# Development and testing
coverage>=7.0.0
factory-boy>=3.3.0

# Production dependencies
gunicorn>=21.0.0
whitenoise>=6.5.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3